Tidying and styling (#16223)
